tf.Datasets and tests
ERP can now create tf.Datasets, network can be trained on them

DQN now actually compares to the correct q_values. the reduce_max needs to be replaced with action

ERP_tests reveal the filling of lists is a little faster. As arrays are more handy later on, I stuck to them.